refactor: rename files to avoid using command when not a base command
