+++ src/commands/data/create/record.ts
@@ -9,7 +9,7 @@ import { Messages } from '@salesforce/core';
 import type { SaveResult } from '@jsforce/jsforce-node';
 import { SfCommand, Flags } from '@salesforce/sf-plugins-core';
 import { orgFlags, perflogFlag } from '../../../flags.js';
-import { stringToDictionary, collectErrorMessages } from '../../../dataCommand.js';
+import { stringToDictionary, collectErrorMessages } from '../../../dataUtils.js';
 
 Messages.importMessagesDirectoryFromMetaUrl(import.meta.url);
 const messages = Messages.loadMessages('@salesforce/plugin-data', 'record.create');
