+++ src/commands/data/delete/record.ts
@@ -9,7 +9,7 @@ import { Messages, SfError } from '@salesforce/core';
 import type { SaveResult } from '@jsforce/jsforce-node';
 import { SfCommand, Flags } from '@salesforce/sf-plugins-core';
 import { orgFlags, perflogFlag } from '../../../flags.js';
-import { collectErrorMessages, query } from '../../../dataCommand.js';
+import { collectErrorMessages, query } from '../../../dataUtils.js';
 
 Messages.importMessagesDirectoryFromMetaUrl(import.meta.url);
 const messages = Messages.loadMessages('@salesforce/plugin-data', 'record.delete');
