+++ src/commands/data/delete/resume.ts
@@ -8,7 +8,7 @@
 import { Messages } from '@salesforce/core';
 import { BulkResultV2 } from '../../../types.js';
 import { BulkDeleteRequestCache } from '../../../bulkDataRequestCache.js';
-import { ResumeBulkCommand } from '../../../resumeBulkCommand.js';
+import { ResumeBulkCommand } from '../../../resumeBulkBaseCommand.js';
 import { isBulkV2RequestDone } from '../../../bulkUtils.js';
 
 Messages.importMessagesDirectoryFromMetaUrl(import.meta.url);
