+++ src/commands/data/update/record.ts
@@ -9,7 +9,7 @@ import { Messages, SfError } from '@salesforce/core';
 import type { SaveError, SaveResult } from '@jsforce/jsforce-node';
 import { SfCommand, Flags } from '@salesforce/sf-plugins-core';
 import { orgFlags } from '../../../flags.js';
-import { collectErrorMessages, query, stringToDictionary } from '../../../dataCommand.js';
+import { collectErrorMessages, query, stringToDictionary } from '../../../dataUtils.js';
 
 Messages.importMessagesDirectoryFromMetaUrl(import.meta.url);
 const messages = Messages.loadMessages('@salesforce/plugin-data', 'record.update');
