+++ src/resumeBulkBaseCommand.ts
